take 3: single index.d.ts file to avoid aliases, successfully exporting Agent
- Back to a single type file, to avoid aliases and the poor dev
  experience of "jump to definition" in an IDE landing on the *alias*
  assignment rather than the actual type definition.
- Using 'interface Agent' rather than 'class Agent' allows exporting it
  in the 'apm' namespace *and* having a singleton of that interface be
  the default export (via 'export = apm').
- Also add test/types/transpile-default/... to test types with a
  TypeScript project *not* using 'esModuleInterop'. We document how to
  use the agent this way
  (https://www.elastic.co/guide/en/apm/agent/nodejs/current/es-modules.html)
  but don't have a test for it that I see.
