+++ package.json
@@ -16,7 +16,7 @@
     "test:deps": "dependency-check start.js index.js 'lib/**/*.js' 'test/**/*.js' --no-dev -i async_hooks -i perf_hooks -i parseurl",
     "test:tav": "tav --quiet",
     "test:docs": "./test/script/docker/run_docs.sh",
-    "test:types": "tsc --project test/types/tsconfig.json && tsc --project test/types/transpile/tsconfig.json && node test/types/transpile/index.js",
+    "test:types": "tsc --project test/types/tsconfig.json && tsc --project test/types/transpile/tsconfig.json && node test/types/transpile/index.js && tsc --project test/types/transpile-default/tsconfig.json && node test/types/transpile-default/index.js",
     "test:babel": "babel test/babel/src.js --out-file test/babel/out.js && node test/babel/out.js",
     "test:esm": "node --experimental-modules test/esm/index.mjs",
     "bench": "./test/benchmarks/scripts/run-benchmarks.sh",
