+++ test/types/commonjs.js
@@ -1,3 +1,6 @@
+// Test that types work for CommonJS code.
+// `tsc` will error out of there is a type conflict.
+
 'use strict'
 
 const agent = require('../../')
