+++ test/types/start.ts
@@ -1,3 +1,6 @@
+// Test the start.d.ts type file by exercising the API in TypeScript.
+// `tsc` will error out of there is a type conflict.
+
 import agent from '../../start'
 import { IncomingMessage, ServerResponse } from 'http'
 import { Socket } from 'net'
