+++ test/types/transpile-default/tsconfig.json
@@ -1,0 +1,10 @@
+{
+  "compilerOptions": {
+    "target": "es5",
+    "module": "commonjs",
+    "strict": true
+  },
+  "files": [
+    "./index.ts"
+  ]
+}
