+++ test/types/transpile/index.ts
@@ -1,7 +1,10 @@
+// Test that `import apm from ...` works for TypeScript using `esModuleInterop`.
+
 import agent from '../../../'
 
 agent.start({
   captureExceptions: false,
   metricsInterval: '0',
-  centralConfig: false
+  centralConfig: false,
+  cloudProvider: 'none'
 })
